Add case-insensitive player filtering
I'm not sure how steam names should be handled (whether they are
case-insensitive or not), so I decided to make this change here rather
than in the `spyparty-rs` crate itself. This led to duplicating the name
comparison code, but until I'm sure of the right way to handle case, I
wanted to leave it out of the raw replay parsing crate.