Add support for "unknown" maps
Occasionally SpyParty will get a new map added and it may take me a
while to update SpyParsey to recognize it.

In order to alleviate some of the frustration, it is now possible to
filter these unknown maps (using `--map unknown`) which will show any
replays which have a map that SpyParsey doesn't recognize. In most
cases, this will simply be the newest map.

Replays containing unknown maps will also show up in the summary output
(as "Unknown" under the "Map" heading).